chore: organize files in folders pages, layouts
