+++ src/main.tsx
@@ -2,14 +2,10 @@ import React from "react";
 import ReactDOM from "react-dom/client";
 import { createBrowserRouter, RouterProvider } from "react-router-dom";
 
-import "./styles/index.css";
+import { Applications, Apply, Eligibility, Home, Rewards } from "./pages";
+import { Layout } from "./layouts";
 
-import { Home } from "./Home";
-import { Eligibility } from "./Eligibility";
-import { Apply } from "./Apply";
-import { Applications } from "./Applications";
-import { Rewards } from "./Rewards";
-import { Layout } from "./Layout";
+import "./styles/index.css";
 
 const router = createBrowserRouter([
   {
@@ -43,5 +39,5 @@ const router = createBrowserRouter([
 ReactDOM.createRoot(document.getElementById("root")!).render(
   <React.StrictMode>
     <RouterProvider router={router} />
-  </React.StrictMode>,
+  </React.StrictMode>
 );
