+++ src/pages/Rewards.tsx
