+++ src/pages/index.ts
@@ -1,0 +1,5 @@
+export { Applications } from "./Applications";
+export { Apply } from "./Apply";
+export { Eligibility } from "./Eligibility";
+export { Home } from "./Home";
+export { Rewards } from "./Rewards";
